# openapi.yaml
...
openapi: 3.1.0
info:
  title: SRE 平台 - API 規格書
  description: |
    本 API 規格為 SRE 平台前後端開發的統一數據契約。

    ## 主要功能模組
    - **身份與存取管理**：用戶、團隊、角色和權限管理
    - **資源管理**：基礎設施資源監控和群組管理
    - **事件管理**：告警規則、靜音規則和事件處理
    - **自動化中心**：腳本管理、排程任務和執行日誌
    - **儀表板**：基礎設施洞察和 SRE 戰情室
    - **通知管理**：通知策略、管道和歷史記錄
    - **分析中心**：容量規劃和性能分析
    - **平台設定**：標籤管理、郵件設定、身份驗證

    它定義了所有資源、事件、自動化、管理等模組的介面，
    並遵循 RESTful 設計原則與專案開發規範。
  version: "3.0.0"
  contact:
    name: SRE 平台開發團隊
    email: sre-platform-dev@detectviz.com
    url: https://github.com/detectviz/sre-platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://detectviz.com/terms
servers:
  - url: /api/v1
    description: 相對路徑 (適用於代理配置)

# ============================================
# 擴展屬性 (Vendor Extensions)
# ============================================
...
paths:
...
  /events:
    get:
      tags: [事件管理 (Event Management)]
      summary: 獲取事件列表
      operationId: listEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          description: 事件狀態篩選
          schema:
            type: string
            enum: [FIRING, RESOLVED, PENDING]
        - name: severity
          in: query
          description: 嚴重度篩選
          schema:
            type: string
            enum: [CRITICAL, WARNING, INFO]
        - name: incident_id
          in: query
          description: 關聯事故ID篩選
          schema:
            type: string
        - name: resource_name
          in: query
          description: 資源名稱關鍵字搜尋
          schema:
            type: string
        - name: source
          in: query
          description: 事件來源篩選
          schema:
            type: string
        - name: sort_by
          in: query
          description: 排序欄位
          schema:
            type: string
        - name: sort_order
          in: query
          description: 排序順序
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: "事件列表"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCollection"
...
  /events/{eventId}/ai-analysis:
    post:
      tags: [事件管理 (Event Management)]
      summary: 觸發 AI 事件分析
      description: "建立 AI 分析報告草稿並非同步執行分析流程。"
      operationId: triggerEventAIAnalysis
      security: [{ bearerAuth: [] }]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIAnalysisRequest"
      responses:
        "202":
          description: 已建立分析報告草稿
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIAnalysisCreationResponse"
        "400": { description: 請求格式錯誤 }
        "500": { description: 伺服器發生錯誤 }
...
  /ai/analysis-reports/{reportId}:
    get:
      tags: [事件管理 (Event Management)]
      summary: 取得 AI 事件分析報告
      description: "依 report_id 查詢非同步分析的最新狀態與內容。"
      operationId: getAIAnalysisReport
      security: [{ bearerAuth: [] }]
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIAnalysisReport"
        "404":
          description: 找不到對應的報告
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    AIAnalysisStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCESS, FAILED]
    AIAnalysisRequest:
      type: object
      properties:
        event_context:
          type: object
          additionalProperties: true
          description: "事件快照或其他提供給 AI 的補充資訊"
    AIAnalysisCreationResponse:
      type: object
      required: [report_id, status]
      properties:
        report_id: { type: string, example: "rep_a1b2c3" }
        status:
          $ref: "#/components/schemas/AIAnalysisStatus"
    AIAnalysisEvidence:
      type: object
      properties:
        type: { type: string }
        description: { type: string }
        link:
          type: object
          nullable: true
          properties:
            name: { type: string }
            url: { type: string, format: uri }
        timestamp:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    AIAnalysisRootCause:
      type: object
      properties:
        text: { type: string }
        confidence_score: { type: number, format: float }
        probable_causes:
          type: array
          items: { type: string }
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/AIAnalysisEvidence"
    AIAnalysisAffectedResource:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        role: { type: string, nullable: true }
    AIAnalysisImpact:
      type: object
      properties:
        text: { type: string }
        affected_resources:
          type: array
          items:
            $ref: "#/components/schemas/AIAnalysisAffectedResource"
        user_impact: { type: string, nullable: true }
        duration_minutes: { type: integer, nullable: true }
        severity: { type: string, nullable: true }
    AIAnalysisRecommendedAction:
      type: object
      properties:
        title: { type: string }
        action_type: { type: string }
        risk: { type: string }
        summary: { type: string, nullable: true }
        action_data:
          type: object
          additionalProperties: true
    AIAnalysisReport:
      type: object
      required: [report_id, event_id, status, created_at, updated_at]
      properties:
        report_id: { type: string }
        event_id: { type: string }
        status:
          $ref: "#/components/schemas/AIAnalysisStatus"
        event_summary: { type: string, nullable: true }
        root_cause_analysis:
          allOf:
            - $ref: "#/components/schemas/AIAnalysisRootCause"
          nullable: true
        impact_assessment:
          allOf:
            - $ref: "#/components/schemas/AIAnalysisImpact"
          nullable: true
        recommended_actions:
          type: array
          items:
            $ref: "#/components/schemas/AIAnalysisRecommendedAction"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/AIAnalysisEvidence"
        error_message: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        raw_llm_response:
          type: object
          nullable: true
          additionalProperties: true
...